:sparkles: feat: post, get and delete cart data from userId

+++ src/components/Cart/index.tsx
@@ -15,21 +15,28 @@ const Cart = () => {
   const { cartItems: items, setCartItems, setError } = useSidebarItems()
   const dispatch = useDispatch()
 
-  const linkReal = 'https://efood-backend.onrender.com'
-  const linkLocal = 'http://localhost:5000'
-
-  const handleDeleteItem = async (id: number) => {
+  const handleDeleteItem = async (dishId: number) => {
     try {
-      const response = await fetch(`${linkLocal}/api/cart/${id}`, {
-        method: 'DELETE'
-      })
+      const userId = Number(localStorage.getItem('userId'))
+
+      if (!userId) {
+        throw new Error('Usuário não autenticado')
+      }
+
+      const response = await fetch(
+        `https://efood-backend.onrender.com/api/cart/${userId}/${dishId}`,
+        {
+          method: 'DELETE'
+        }
+      )
 
       if (!response.ok) {
         throw new Error('Erro ao remover item do carrinho')
       }
 
-      const updatedCart = await response.json() // Adiciona esta linha para obter os itens atualizados da resposta da API
-      setCartItems(updatedCart.cartItems) // Atualiza os itens com a resposta da API
+      const updatedCart = await response.json()
+
+      setCartItems([...updatedCart.items])
     } catch (error: any) {
       console.error('Erro ao deletar item:', error.message)
       setError(error.message)
@@ -49,7 +56,11 @@ const Cart = () => {
     <>
       <ul>
         {items.map((item) => (
-          <CartItem key={item.id} item={item} onDelete={handleDeleteItem} />
+          <CartItem
+            key={item.id}
+            item={item}
+            onDelete={() => handleDeleteItem(item.id)}
+          />
         ))}
       </ul>
       <S.PriceContainer>
